refactor(game): move world-meta reader to data-access/world-meta
Moves the pure `defaultedWorldMeta` helper and the read-only `getWorldMetaServer`
fetch out of data-server into lib/game/data-access/world-meta.ts. data-server's
transactional `readWorldMetaInTx` imports `defaultedWorldMeta` back, and
`getWorldMetaServer` is re-exported from data-server so existing
`@/lib/game/data-server` importers are unaffected.

No cycle: world-meta.ts depends only on ../types + data-access/{collections,db}.

Behavior-preserving. Verified: game suite 1460/1460, tsc + eslint clean.
data-server.ts 7,444 → 7,416.

Stacked on the data-access chain (#1571 → #1570 → #1569 → #1567).
